Added partial code sequence for initializing display capsules
